defmt_decoder: Move `fn check_version` into `mod elf2table`
